requirements.txt
# Core ML/DL libraries
tensorflow>=2.13.0
keras>=2.13.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration and utilities
PyYAML>=6.0
tqdm>=4.65.0

# Web interface (optional)
streamlit>=1.25.0

# Jupyter notebook support
jupyter>=1.0.0
ipywidgets>=8.0.0

# Data augmentation
albumentations>=1.3.0

# Model optimization
tensorflow-model-optimization>=0.7.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Kaggle integration
kagglehub>=0.1.0

# Additional utilities
requests>=2.31.0
python-dotenv>=1.0.0
